vendor VapourSynth4 SDK and AviSynth+ headers + build fixes
- Vendor VapourSynth4.h from upstream (git master, API 4.x)
- Vendor AviSynth+ headers (avisynth.h, avs/{config,capi,types,cpuid,posix}.h)
- Update CMakeLists.txt to prefer vendored headers over system paths
- Fix narrowing errors: explicit static_cast<uint64_t>() for VideoInfo::CS_* constants
- Fix missing validate() function in AvsReader.h
- Fix clang warnings: remove 'static' from explicit template specializations

+++ AvsReader.cpp
@@ -75,14 +75,14 @@ get_vs_video_format(VSVideoFormat* fmt, int pixel_type, int bitdepth,
         int subW;
         int subH;
     } table[] = {
-        { VideoInfo::CS_BGR32, cfRGB, 0, 0 },
-        { VideoInfo::CS_BGR24, cfRGB, 0, 0 },
-        { VideoInfo::CS_YV24,  cfYUV, 0, 0 },
-        { VideoInfo::CS_YV16,  cfYUV, 1, 0 },
-        { VideoInfo::CS_YV411, cfYUV, 2, 0 },
-        { VideoInfo::CS_I420,  cfYUV, 1, 1 },
-        { VideoInfo::CS_YV12,  cfYUV, 1, 1 },
-        { VideoInfo::CS_Y8,    cfGray, 0, 0 },
+        { static_cast<uint64_t>(VideoInfo::CS_BGR32), cfRGB, 0, 0 },
+        { static_cast<uint64_t>(VideoInfo::CS_BGR24), cfRGB, 0, 0 },
+        { static_cast<uint64_t>(VideoInfo::CS_YV24),  cfYUV, 0, 0 },
+        { static_cast<uint64_t>(VideoInfo::CS_YV16),  cfYUV, 1, 0 },
+        { static_cast<uint64_t>(VideoInfo::CS_YV411), cfYUV, 2, 0 },
+        { static_cast<uint64_t>(VideoInfo::CS_I420),  cfYUV, 1, 1 },
+        { static_cast<uint64_t>(VideoInfo::CS_YV12),  cfYUV, 1, 1 },
+        { static_cast<uint64_t>(VideoInfo::CS_Y8),    cfGray, 0, 0 },
         { 0, 0, 0, 0 }  /* sentinel */
     };
